Check for authorized user of slice

+++ dspace/src/de/zib/gndms/dspace/service/SliceServiceImpl.java
@@ -16,6 +16,7 @@ package de.zib.gndms.dspace.service;
  * limitations under the License.
  */
 
+import com.sun.servicetag.UnauthorizedAccessException;
 import de.zib.gndms.common.dspace.service.SliceService;
 import de.zib.gndms.common.logic.config.Configuration;
 import de.zib.gndms.common.model.FileStats;
@@ -110,8 +111,9 @@ public class SliceServiceImpl implements SliceService {
 		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKindId, sliceId, dn );
 
         try {
-            // check for the existance of that slice
+            // check for the existence of that slice
             findSliceOfKind( subspaceId, sliceKindId, sliceId );
+            checkOwner( dn, subspaceId, sliceId );
             
             return new ResponseEntity< Facets >( new Facets( listFacetsOfSlice( subspaceId, sliceKindId, sliceId ) ), headers, HttpStatus.OK );
         } catch ( NoSuchElementException ne ) {
@@ -122,19 +124,20 @@ public class SliceServiceImpl implements SliceService {
         }
 	}
 
-	@Override
+    @Override
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{slice}/config", method = RequestMethod.PUT)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Void> setSliceConfiguration(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
-			@PathVariable final String slice,
+			@PathVariable final String sliceId,
 			@RequestBody final Configuration config,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, slice, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn);
 
 		try {
-			Slice slic = findSliceOfKind(subspace, sliceKind, slice);
+			Slice slic = findSliceOfKind( subspaceId, sliceKind, sliceId );
+            checkOwner( dn, subspaceId, sliceId );
 
 			SliceConfiguration slConfig = SliceConfiguration
 					.checkSliceConfig(config);
@@ -159,18 +162,20 @@ public class SliceServiceImpl implements SliceService {
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{slice}", method = RequestMethod.POST)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Specifier<Void>> transformSlice(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
-			@PathVariable final String slice,
+			@PathVariable final String sliceId,
 			@RequestBody final Specifier<Void> newSliceKind,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, slice, dn);
+		GNDMSResponseHeader headers = setHeaders(subspaceId, sliceKind, sliceId, dn);
 
 		try {
-			Slice slic = findSliceOfKind(subspace, sliceKind, slice);
-			SliceKind newSliceK = sliceKindProvider.get(subspace,
+			Slice slic = findSliceOfKind( subspaceId, sliceKind, sliceId );
+            checkOwner( dn, subspaceId, sliceId );
+
+			SliceKind newSliceK = sliceKindProvider.get( subspaceId,
                     newSliceKind.getUrl());
-			Subspace space = subspaceProvider.get(subspace);
+			Subspace space = subspaceProvider.get( subspaceId );
 
             EntityManager em = emf.createEntityManager();
 			TxFrame tx = new TxFrame( em );
@@ -180,7 +185,7 @@ public class SliceServiceImpl implements SliceService {
 						dn, slic.getTerminationTime(),
 						newSliceK, space, slic.getTotalStorageSize(), null);
 				action.setOwnEntityManager( em );
-				logger.info("Calling action for transforming sliceId " + slice
+				logger.info("Calling action for transforming sliceId " + sliceId
 						+ ".");
 				action.call();
 				tx.commit();
@@ -195,9 +200,9 @@ public class SliceServiceImpl implements SliceService {
 
 			HashMap<String, String> urimap = new HashMap<String, String>(2);
 			urimap.put("service", "dspace");
-			urimap.put(UriFactory.SUBSPACE, subspace);
-			urimap.put(UriFactory.SLICE_KIND, sliceKind);
-			urimap.put(UriFactory.SLICE, slice);
+			urimap.put(UriFactory.SUBSPACE, subspaceId );
+			urimap.put(UriFactory.SLICE_KIND, sliceKind );
+			urimap.put(UriFactory.SLICE, sliceId );
 			spec.setUriMap(new HashMap<String, String>(urimap));
 			spec.setUrl(uriFactory.quoteUri(urimap));
 
@@ -221,6 +226,8 @@ public class SliceServiceImpl implements SliceService {
 		GNDMSResponseHeader headers = setHeaders(subspaceId, sliceKindId, sliceId, dn);
 
 		try {
+            checkOwner( dn, subspaceId, sliceId );
+
             // submit action
             final Taskling ling = sliceProvider.deleteSlice(subspaceId, sliceId);
 
@@ -249,6 +256,8 @@ public class SliceServiceImpl implements SliceService {
 		final GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKindId, sliceId, dn );
 
 		try {
+            checkOwner( dn, subspaceId, sliceId );
+
 			final Subspace space = subspaceProvider.get( subspaceId );
 			final Slice slice = findSliceOfKind( subspaceId, sliceKindId, sliceId );
 			final String path = space.getPathForSlice( slice );
@@ -270,16 +279,18 @@ public class SliceServiceImpl implements SliceService {
     @RequestMapping(value = "/_{subspace}/_{sliceKind}/_{sliceId}/files", method = RequestMethod.POST)
     @Secured( "ROLE_USER" )
     public ResponseEntity<Void> setFileContents(
-            @PathVariable final String subspace,
+            @PathVariable final String subspaceId,
             @PathVariable final String sliceKind,
             @PathVariable final String sliceId,
             @RequestParam( "files" ) final List< MultipartFile > files,
             @RequestHeader("DN") final String dn) {
-        GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, sliceId, dn);
+        GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn);
 
         try {
-            Subspace space = subspaceProvider.get(subspace);
-            Slice slice = findSliceOfKind(subspace, sliceKind, sliceId);
+            checkOwner( dn, subspaceId, sliceId );
+
+            Subspace space = subspaceProvider.get( subspaceId );
+            Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId );
             String path = space.getPathForSlice(slice);
             
             for( MultipartFile file: files ) {
@@ -308,16 +319,18 @@ public class SliceServiceImpl implements SliceService {
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{slice}/files", method = RequestMethod.DELETE)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Void> deleteFiles(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
-			@PathVariable final String slice,
+			@PathVariable final String sliceId,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, slice, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn);
 
 		try {
-			Subspace space = subspaceProvider.get(subspace);
-			Slice slic = findSliceOfKind(subspace, sliceKind, slice);
-			String path = space.getPathForSlice(slic);
+            checkOwner( dn, subspaceId, sliceId );
+
+			Subspace space = subspaceProvider.get( subspaceId );
+			Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId );
+			String path = space.getPathForSlice( slice );
 
             File f = new File( path );
             String[] fl = f.list( );
@@ -339,16 +352,19 @@ public class SliceServiceImpl implements SliceService {
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{slice}/gsiftp", method = RequestMethod.GET)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<String> getGridFtpUrl(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
-			@PathVariable final String slice,
+			@PathVariable final String sliceId,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, slice, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn );
+
 		try {
-			Subspace space = subspaceProvider.get(subspace);
-			Slice slic = findSliceOfKind(subspace, sliceKind, slice);
+            checkOwner( dn, subspaceId, sliceId );
+
+			Subspace space = subspaceProvider.get( subspaceId );
+			Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId );
 			return new ResponseEntity<String>(
-					space.getGsiFtpPathForSlice(slic), headers, HttpStatus.OK);
+					space.getGsiFtpPathForSlice( slice ), headers, HttpStatus.OK);
 		} catch (NoSuchElementException ne) {
 			logger.warn(ne.getMessage());
 			return new ResponseEntity<String>(null, headers,
@@ -361,16 +377,19 @@ public class SliceServiceImpl implements SliceService {
             method = RequestMethod.GET)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Void> listFileContent(
-            @PathVariable final String subspace,
+            @PathVariable final String subspaceId,
             @PathVariable final String sliceKind,
             @PathVariable final String sliceId,
             @PathVariable final String fileName,
             @RequestParam( value="attrs", required = false) final List<String> attrs,
             @RequestHeader("DN") final String dn, final OutputStream out) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, sliceId, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn );
+
 		try {
-			Subspace space = subspaceProvider.get(subspace);
-			Slice slice = findSliceOfKind(subspace, sliceKind, sliceId);
+            checkOwner( dn, subspaceId, sliceId );
+
+			Subspace space = subspaceProvider.get( subspaceId );
+			Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId);
 			String path = space.getPathForSlice(slice);
 			File file = new File(path + File.separatorChar + fileName);
 
@@ -414,17 +433,19 @@ public class SliceServiceImpl implements SliceService {
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{sliceId}/_{fileName}", method = RequestMethod.POST)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Void> setFileContent(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
 			@PathVariable final String sliceId,
 			@PathVariable final String fileName,
 			@RequestParam( "file" ) final MultipartFile file,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, sliceId, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn );
 
 		try {
-			Subspace space = subspaceProvider.get(subspace);
-			Slice slice = findSliceOfKind(subspace, sliceKind, sliceId);
+            checkOwner( dn, subspaceId, sliceId );
+
+			Subspace space = subspaceProvider.get( subspaceId );
+			Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId );
 			String path = space.getPathForSlice(slice);
 			File newFile = new File(path + File.separatorChar + fileName);
 
@@ -455,17 +476,19 @@ public class SliceServiceImpl implements SliceService {
 	@RequestMapping(value = "/_{subspace}/_{sliceKind}/_{slice}/_{fileName}", method = RequestMethod.DELETE)
     @Secured( "ROLE_USER" )
 	public ResponseEntity<Void> deleteFile(
-			@PathVariable final String subspace,
+			@PathVariable final String subspaceId,
 			@PathVariable final String sliceKind,
-			@PathVariable final String slice,
+			@PathVariable final String sliceId,
 			@PathVariable final String fileName,
 			@RequestHeader("DN") final String dn) {
-		GNDMSResponseHeader headers = setHeaders(subspace, sliceKind, slice, dn);
+		GNDMSResponseHeader headers = setHeaders( subspaceId, sliceKind, sliceId, dn );
 
 		try {
-			Subspace space = subspaceProvider.get(subspace);
-			Slice slic = findSliceOfKind(subspace, sliceKind, slice);
-			String path = space.getPathForSlice(slic);
+            checkOwner( dn, subspaceId, sliceId );
+
+			Subspace space = subspaceProvider.get( subspaceId );
+			Slice slice = findSliceOfKind( subspaceId, sliceKind, sliceId );
+			String path = space.getPathForSlice( slice );
 
             if( directoryAux.deleteDirectory( dn, path ) ) {
                 return new ResponseEntity< Void >( null, headers, HttpStatus.OK );
@@ -543,11 +566,21 @@ public class SliceServiceImpl implements SliceService {
 		SliceKind sliceK = sliceKindProvider.get( subspaceId, sliceKindId );
 
 		if( !slice.getKind().equals( sliceK ) ) {
-            logger.error( "Slice " + sliceId + " is of sliceKind " + slice.getKind().getId() + " instead of " + sliceKindId );
+            logger.error( "Slice " + sliceId + " is of sliceKind " +
+                    slice.getKind().getId() + " instead of " + sliceKindId );
 			throw new NoSuchElementException();
 		}
 		return slice;
 	}
+
+    private void checkOwner( String dn, String subspaceId, String sliceId ) throws NoSuchElementException {
+        Slice slice = sliceProvider.getSlice( subspaceId, sliceId );
+        
+        if( ! slice.getOwner().equals( dn ) ) {
+            logger.debug( "User " + dn + " tried to access slice " + sliceId + ", owned by " + slice.getOwner() + "." );
+            throw new UnauthorizedAccessException( "User " + dn + " does not own slice " + sliceId + "." );
+        }
+    }
 
     void recursiveListFiles( String path, String prefix, List<FileStats> list ) {
         List< String > flatContents = directoryAux.listContent( path );
